animacio

+++ src/motor/AbsztraktJatek.java
@@ -1,7 +1,7 @@
 package motor;
 public abstract class AbsztraktJatek {
 
-    public abstract void frissit(JatekMag jm,float dt);
+    public abstract void frissit(JatekMag jm,double dt);
     public abstract void render(JatekMag jm,Render r);
     
 }
